requirements.txt
fastapi
uvicorn[standard]
openai>=2.7.1
psycopg2-binary
python-dotenv
gunicorn
pydantic
pydantic-settings
email-validator
urllib3
SQLAlchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
openpyxl
redis>=5.0.0

# Development tools
ruff

# Testing dependencies
pytest
pytest-asyncio
pytest-cov
pytest-mock
httpx
